requirements.txt
attrs==23.2.0
aws-cdk-lib==2.147.3
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
aws-lambda-powertools==2.40.1
boto3==1.34.136
boto3-stubs==1.34.137
botocore==1.34.136
botocore-stubs==1.34.137
cattrs==23.2.3
certifi==2024.6.2
charset-normalizer==3.3.2
cloudpickle==2.2.1
constructs==10.3.0
contourpy==1.1.1
cycler==0.12.1
dill==0.3.8
docker==7.1.0
exceptiongroup==1.2.1
fonttools==4.53.0
google-pasta==0.2.0
haversine==2.8.1
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
jmespath==1.0.1
joblib==1.4.2
jsii==1.101.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llvmlite==0.41.1
matplotlib==3.7.5
multiprocess==0.70.16
mypy-boto3-sagemaker-runtime==1.34.0
numba==0.58.1
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pathos==0.3.2
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pox==0.3.4
ppft==1.7.6.8
protobuf==4.25.3
psutil==6.0.0
psycopg2==2.9.9
publication==0.0.3
pyod==2.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
s3transfer==0.10.2
sagemaker==2.224.2
schema==0.7.7
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
smdebug-rulesconfig==1.0.1
SQLAlchemy==2.0.31
tblib==3.0.0
threadpoolctl==3.5.0
tqdm==4.66.4
typeguard==2.13.3
types-awscrt==0.20.12
types-s3transfer==0.10.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
zipp==3.19.2
